requirements.txt
tensorboard==2.2.2
torch==1.7.0 #note: please use cuda 11.0, conda install pytorch==1.7.0 torchvision==0.8.0 torchaudio==0.7.0 cudatoolkit=11.0 -c pytorch -y
torchvision==0.8.0
numpy>=1.21
scikit-image
scikit-learn==1.0.2 #note: please use python 3.7 or 3.6
scipy==1.4.1
seaborn==0.9.0
pandas==0.24.2
Pillow>=9.0.1
h5py
metric-learn
tqdm
scikit-learn-intelex
# please run conda install -c pytorch -c nvidia faiss-gpu -y for --fast_kmeans